[fix] tmphomeni に変更しhome　gameghomeをhomeに
