+++ spa/views/Home.js
@@ -1,31 +1,41 @@
-import Footer from "../components/Footer.js";
-import Header from "../components/Header.js";
-import stateManager from "../stateManager.js";
+import SPA from "../spa.js";
 
 export default function Home() {
   return `
-        ${Header({ title: "ホーム" })}
-        <p>これはホームページです。</p>
-        <img src="./assets/42.png" alt="ロゴ">
-        <button onclick="SPA.navigate('/store')">storeへ</button>
-        <button onclick="SPA.navigate('/api')">APIへ</button>
-        <button onclick="SPA.navigate('/404')">404へ</button>
-        <button onclick="SPA.navigate('/profile')">profileへ</button>
-        <button onclick="SPA.navigate('/history/match')">Match Historyへ</button>
-        <button onclick="SPA.navigate('/signup')">signupへ</button>
-        <button onclick="SPA.navigate('/login')">loginへ</button>
-        <button onclick="SPA.navigate('/profile/username')">changeUsernameへ</button>
-        <button onclick="SPA.navigate('/profile/mail')">changeMailへ</button>
-        <button onclick="SPA.navigate('/profile/password')">changePasswordへ</button>
-        <button onclick="SPA.navigate('/profile/avatar')">changeAvatarへ</button>
-
-        <button onclick="SPA.navigate('/game')">gameへ</button>
-        <button onclick="SPA.navigate('/game/setup')">setupgameへ</button>
-        <button onclick="SPA.navigate('/game/result')">gameresultへ</button>
-        <button onclick="SPA.navigate('/matching/quick-play')">QuickPlayMatchingへ</button>
-        <button onclick="SPA.navigate('/matching/tournament')">TournamentMatchingへ</button>
-        ${Footer({ text: "© 2025 My Company" })}
-        <div> store sample</div>
-        ${stateManager.state.count}
+        <img src="/assets/user.png" class="profile-icon js-btn-profile">
+        <img src="/assets/friend.png" class="friend-icon js-btn-friend">
+        <div class="d-flex vh-100 align-items-center justify-content-center text-center">
+            <div class="d-grid gap-4 col-4 mx-auto">
+                <button class="btn btn-primary btn-quickplay js-btn-quickplay" type="button">
+                  <img src="/assets/table_tennis.png" class="icon-table-tennis">
+                  Quick Play
+                </button>
+                <button class="btn btn-primary btn-tournament js-btn-tournament" type="button">
+                  <img src="/assets/tournament.png" class="icon-tournament">
+                  Tournament
+                </button>
+            </div>
+        </div>
     `;
+}
+
+export function setupHome() {
+  const profileButton = document.querySelector(".js-btn-profile");
+  const friendButton = document.querySelector(".js-btn-friend");
+
+  const quickplayButton = document.querySelector(".js-btn-quickplay");
+  const tournamentButton = document.querySelector(".js-btn-tournament");
+
+  profileButton.addEventListener("click", () => {
+    SPA.navigate("/profile");
+  });
+  friendButton.addEventListener("click", () => {
+    SPA.navigate("/friend");
+  });
+  quickplayButton.addEventListener("click", () => {
+    SPA.navigate("/matching/quick-play");
+  });
+  tournamentButton.addEventListener("click", () => {
+    SPA.navigate("/matching/tournament");
+  });
 }
